# webui/bff/assets/swagger/beemon.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/beemon.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/namespaces/{nsType}/{nsInode}": {
      "get": {
        "operationId": "BeemonService_GetNamespaceDetails",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetNamespaceDetailsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "nsType",
            "description": "\"mnt\", \"net\", \"uts\", \"user\"",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "nsInode",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "referencePid",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "BeemonService"
        ]
      }
    },
    "/api/v1/processes": {
      "get": {
        "operationId": "BeemonService_ListProcesses",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListProcessesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "filterName",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "BeemonService"
        ]
      }
    },
    "/api/v1/processes/{pid}/events": {
      "get": {
        "operationId": "BeemonService_StreamEvents",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1EventBatch"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1EventBatch"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "BeemonService"
        ]
      }
    },
    "/api/v1/processes/{pid}/metadata": {
      "get": {
        "operationId": "BeemonService_GetProcessMetadata",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetProcessMetadataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "BeemonService"
        ]
      }
    },
    "/api/v1/processes/{pid}/network_flows": {
      "get": {
        "operationId": "BeemonService_GetNetworkFlows",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetNetworkFlowsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "BeemonService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1EventBatch": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Event"
          }
        }
      }
    },
...
    "v1GetNamespaceDetailsResponse": {
      "type": "object",
      "properties": {
        "nsType": {
          "type": "string"
        },
        "nsInode": {
          "type": "string"
        },
        "mountInfo": {
          "type": "string"
        },
        "netLinks": {
          "type": "string"
        },
        "netRoutes": {
          "type": "string"
        },
        "utsInfo": {
          "type": "string"
        },
        "userMaps": {
          "type": "string"
        }
      }
    },
    "v1GetNetworkFlowsResponse": {
      "type": "object",
      "properties": {
        "flows": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1NetworkFlow"
          }
        }
      }
    },
    "v1GetProcessMetadataResponse": {
      "type": "object",
      "properties": {
        "process": {
          "$ref": "#/definitions/v1Process"
        },
        "parent": {
          "$ref": "#/definitions/v1Process"
        },
        "children": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Process"
          }
        },
        "hostNamespaces": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "v1ListProcessesResponse": {
      "type": "object",
      "properties": {
        "processes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Process"
          }
        },
        "hostMemoryTotalBytes": {
          "type": "string",
          "format": "uint64"
        },
        "hostNamespaces": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "hostCpuPerCorePercent": {
          "type": "array",
          "items": {
            "type": "number",
            "format": "float"
          }
        },
        "hostIoReadBytesPerSec": {
          "type": "string",
          "format": "uint64",
          "description": "Host-wide per-second I/O totals, summed from the eBPF process_io_stats map.\nSame source as per-process io_*_bytes fields, but aggregated across all PIDs\nin the snapshot to give a single dashboard gauge. 0 means the BPF map is cold."
        },
        "hostIoWriteBytesPerSec": {
          "type": "string",
          "format": "uint64"
        },
        "hostNetRxBytesPerSec": {
          "type": "string",
          "format": "uint64"
        },
        "hostNetTxBytesPerSec": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
    "v1NetworkFlow": {
      "type": "object",
      "properties": {
        "localAddress": {
          "type": "string"
        },
        "remoteAddress": {
          "type": "string"
        },
        "localPort": {
          "type": "integer",
          "format": "int64"
        },
        "remotePort": {
          "type": "integer",
          "format": "int64"
        },
        "protocol": {
          "type": "string",
          "title": "\"TCP\", \"UDP\""
        },
        "rxBytes": {
          "type": "string",
          "format": "uint64"
        },
        "txBytes": {
          "type": "string",
          "format": "uint64"
        },
        "rxPackets": {
          "type": "string",
          "format": "uint64"
        },
        "txPackets": {
          "type": "string",
          "format": "uint64"
        },
        "dnsQuery": {
          "type": "string"
        }
      }
    },
    "v1OpenFile": {
      "type": "object",
      "properties": {
        "fd": {
          "type": "integer",
          "format": "int64"
        },
        "path": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "title": "e.g., \"socket\", \"pipe\", \"regular\", \"directory\", \"char\", \"fifo\", \"other\""
        }
      }
    },
...
    "v1Process": {
      "type": "object",
      "properties": {
        "pid": {
          "type": "integer",
          "format": "int64"
        },
        "ppid": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "memoryUsageBytes": {
          "type": "string",
          "format": "uint64",
          "title": "Usage"
        },
        "cpuUsagePercent": {
          "type": "number",
          "format": "float"
        },
        "memoryLimitBytes": {
          "type": "string",
          "format": "uint64",
          "description": "0 means no limit",
          "title": "Limits"
        },
        "cpuQuotaUs": {
          "type": "string",
          "format": "uint64",
          "title": "0 means no limit"
        },
        "cpuPeriodUs": {
          "type": "string",
          "format": "uint64"
        },
        "pidsLimit": {
          "type": "string",
          "format": "uint64",
          "title": "0 means no limit (or max)"
        },
        "namespaces": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Namespaces"
        },
        "openFiles": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1OpenFile"
          },
          "title": "Open Files"
        },
        "activeConnections": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1NetworkConnection"
          },
          "title": "Active Network Connections"
        },
        "ioReadBytes": {
          "type": "string",
          "format": "uint64",
          "title": "I/O Stats"
        },
        "ioWriteBytes": {
          "type": "string",
          "format": "uint64"
        },
        "netRxBytes": {
          "type": "string",
          "format": "uint64"
        },
        "netTxBytes": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
  }
}